Переместить функционал, связанный с паролями в core/authentication.py
